making transform and warehouse are loosely coupled. deleting warehouse will remove only the cli profile block
